+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 * Fix cannot create a custom title. (UIEH-1103)
 * Fix Settings assign/usassign user permission. (UIEH-1106)
 * Update permission name for eholdings module. (UIEH-1104)
+* Provider: Re-order accordions display. (UIEH-1044)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)